fix 503 after first subscribe request
The SubscribeActor was shutting itself down after
the first messages and thus wasn't available for
any subsequent requests. This actor isn't really
needed and just added complexity so this has been
folded into the SubscribeApi.
