+++ atlas-lwcapi/src/main/resources/reference.conf
@@ -6,13 +6,6 @@ atlas {
   }
 
   akka {
-    actors = ${?atlas.akka.actors} [
-      {
-        name = "lwc.subscribe"
-        class = "com.netflix.atlas.lwcapi.SubscribeActor"
-      }
-    ]
-
     api-endpoints = ${?atlas.akka.api-endpoints} [
       "com.netflix.atlas.lwcapi.EvaluateApi",
       "com.netflix.atlas.lwcapi.ExpressionApi",
